modified:   contracts/Pixelmap.sol 	modified:   frontend/package-lock.json 	modified:   frontend/package.json 	deleted:    frontend/src/artifacts/build-info/610a417ae3c5b53305833092a45469e9.json 	new file:   frontend/src/artifacts/build-info/689b1b23f8fb022901e72c73605063d0.json 	modified:   frontend/src/artifacts/contracts/Pixelmap.sol/CanvasCollection.dbg.json 	modified:   frontend/src/artifacts/contracts/Pixelmap.sol/Pixelmap.dbg.json 	modified:   frontend/src/artifacts/contracts/Pixelmap.sol/Pixelmap.json 	modified:   frontend/src/components/Art.jsx 	modified:   frontend/src/components/Canvas.jsx 	modified:   frontend/src/components/MobilePortfolio.jsx 	modified:   frontend/src/components/NavBar.jsx 	modified:   frontend/src/hooks/contractAddr.jsx 	modified:   frontend/src/hooks/currencyAddr.jsx 	modified:   frontend/src/hooks/nftContractAddr.jsx 	modified:   scripts/getWETH.js 	modified:   test/pixelmapTest.js

+++ contracts/Pixelmap.sol
@@ -138,7 +138,7 @@ contract Pixelmap is ReentrancyGuard {
                 window[_x][_y].color = _color;
                 window[_x][_y].shapeID = _shape;
 
-                uint pixelID = _x + maxWidth * _y;
+                uint256 pixelID = _x + maxWidth * _y;
                 
                 pixelSVGs[pixelID] = generateShapeSVG(window[_x][_y].shapeID, _x * 9, _y * 9, window[_x][_y].color);
                 uniqueRows[_y] = true;
